+++ src/src/user/user.controller.ts
@@ -23,6 +23,7 @@ import { HasuraService } from '../hasura/hasura.service';
 import { CreateUserDto } from '../helper/dto/create-user.dto';
 import { RegisterFacilitatorDto } from '../helper/dto/register-facilitator.dto';
 import { UserService } from './user.service';
+import { UserV2Service } from './user.v2.service';
 
 @UseInterceptors(SentryInterceptor)
 @Controller('/users')
@@ -32,6 +33,7 @@ export class UserController {
 		private readonly httpService: HttpService,
 		public hasuraService: HasuraService,
 		public userService: UserService,
+		public userV2Service: UserV2Service,
 	) {}
 
 	@Get('/qualification')
@@ -195,6 +197,6 @@ export class UserController {
 		@Res()response:any,
 		@Body()body:any
 	){
-		return this.userService.v2_ifUserExist(role,body,response);
+		return this.userV2Service.isUserExist(role,body,response);
 	}
 }
